backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
redis==5.0.5
python-jose==3.3.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
prometheus-client==0.20.0
sentry-sdk==2.6.0
httpx==0.27.0
loguru==0.7.2
pytest==8.2.0
pytest-asyncio==0.23.6
